+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 3.10.0
+* Add `dataOnlyRequested` to `ThreeDSecureLookupRequest`
+* Add `paymentReaderCardDetails` to `TransactionCreditCardRequest`
+* Add `chargebackProtectionLevel` to `Dispute` and `DisputeSearchRequest`
+* Add `skipAdvancedFraudChecking` to `CreditCardOptionsRequest` and `PaymentMethodOptionsRequest`
+
 ## 3.9.0
 * Add `getPaypalMessages` to `Dispute`
 * Add webhook sample for `GrantedPaymentMethodRevoked`